// src/utils/notion.types.ts
export type NameProp = {
	type: 'title';
	id: string;
	title: Array<TextRichTextItemResponse>;
};

export type DateProp = {
	type: 'date';
	date: {
		start: string;
		end: string | null;
		time_zone: TimeZoneRequest | null;
	};
	id: string;
};